Add named styles for markup languages
For Markdown and Text2Tags the named styles mapping were not good and
there was no named styles in the color schemes for this type of
language. This commit adds common named styles to the color schemes for
them to use and updates the filetypes.{markdown,txt2tags} to use them.