Respect "use_native_windows_dialogs" for Prefs->Interface fonts
When using native dialogs, it overrides the GtkFontButton from opening
a font chooser dialog and instead shows a native win32 font dialog.

Adds a function "win32_show_font_dialog2" that returns the selected
font information and also takes the initial font to select in the
dialog as a parameter.

Adds an enum to prefs.c to replace 1,2,3 for which font to change
in the preferences dialog.