Make app->project->base_path always absolute
Because app->project->base_path can be relative, its direct use
leads to resolving the directory which is relative to the directory
where Geany was started instead of the project directory. Making it
always absolute fixes the problem. The original path is stored under
GeanyProjectPrivate.

In addition, get rid of project_get_base_path() which is not needed now.