Add plugin API functions to manage arbitrary plugin data
Adds new functions plugin_set_field(), plugin_get_field() and
plugin_remove_field() which allow plugins to manage the private
hash table for the plugin. Plugins can set a free_func to have to
pointer freed automatically when the field is overwritten or when
the plugin is removed. plugin_remove_field() allows removing a
pointer from the hash table without calling the free_func on it.