Move long-line/edge-mode setting into GtkScintilla
Also add some enumerations to make the code more readable. Some
functions added in editor.c to map the weird gint (now enum) that
GeanyEditor had, which for whatever reason doesn't map to Scintilla
EDGE_* values, to the GeanyScintillaEdgeMode enum.

TODO: Get rid of the redundant GeanyEditorLongLineSetting enum that
was added to replace plain gints+comments. Changing it will change the
meaning of what's stored in the user's config file, which is why I left
some of this related code.