Add "text" property and related functions to GeanyScintilla
Making the similar functions in sciwrappers use the new functions
but not all other code updated as the sciwrappers functions are
widely used, will check into it as another commit.

TODO: profile the code to make sure SCI_GETCHARACTERPOINTER and
SCI_GETRANGEPOINTER is not too slow (due to moving buffer gap) for
how these functions are used in Geany.