Replace uses of sci_get_length with geany_scintilla_get_text_length
It's a bit longer to type but mostly because currently the
GEANY_SCINTILLA() cast is needed. Eventually editor will have a
GeanyScintilla object so the casts won't be needed.

Also started moving sciwrapper.h functions to top. These are ones
that are part of the plugin API and we need to decide whether to
deprecate them once the equivalent GeanyScintilla API is exposed.