Export `documents_array` using G_MODULE_EXPORT
It is the same on *nix, but will make sure the plugins can see this
symbol on Windows when expanded through some preprocessor macros.
Previously it was expanded from the macros to do (through a number of
other macro expansions), `geany_data->documents_array` instead of
accessing it directly.

Also remove the `GEANY` macro in `geany.h` since this is what that
was used for and is no longer needed/used in core.