Use linker to export build functions
This is a prototype for doing this for all API functions. It would
make the access to private functions impossible since they won't be
available by an include and should ensure that the public symbols are
exported on Windows. It also removes having to have these functions
be specified to be public/API functions. This would probably require
bumping the plugin ABI since it's pulling members off of the end of
the public GeanyFunctions member (TODO: confirm this).