Refactor msgwindow_visible code to use GtkAction
* Remove use of ignore_callback and ui_prefs.msgwindow_visible globals
and use the action instead.
* Use the same action for both the View->Show Message Window and
the menu item in the message window's context menu.
* Rename msgwin_show_hide() to msgwin_set_visible() and add
msgwin_get_visible() to get the visibility of the message window
instead of also using the ui_prefs global for this.
* Don't choose "use action appearance" for the "Show Message Window"
menu item even though it'd avoid some duplication with the action,
since it won't show the accel key hint in the menu unless the
keybindings stuff is changed to use the action's accelerator.
* Add hack  to ui_utils/builder code for the known (and apparently
fixed in newer versions) Glade/GtkBuilder bug related to GtkAccelGroup
for the actions.