Refactor fullscreen code to use GtkAction
* Remove use of ignore_callback and ui_prefs.fullscreen globals and
use the action instead.
* Remove extra fullscreen related code or move it to ui_utils.
* Add "fullscreen" argument to ui_set_fullscreen() and add
ui_get_fullscreen() function and make both operate on the action.
* Remove on_window_state_event handler since we don't really care
whether the window manager actually set fullscreen mode or not, it
either supports it or not.
* Remove window size/fullscreen hack from configuration_apply_settings()
and add a different hack with same effect to ui_set_fullscreen(FALSE).
In both cases when leaving fullscreen, the window is left maximized
(tested with XFWM4).