Mark all plugin API functions to have "default" (public) visibility
Adds a new header `pluginexport.h` to put the macros in, could be
moved into an existing header (support.h?) by I didn't want to drag
a bunch of existing stuff into the source files for this one macro.

Scintilla and TagManager have relative includes, this could be fixed
by changing the include directories of them or, especially important
for Scintilla to avoid upstream changes, to put the marked symbol
declarations in a sparate file (ex. sciwrappers.h could contain another
declaration for each with the visibility macro applied.